+++ internal/logging/mock_logger_test.go
@@ -1,0 +1,8 @@
+package logging
+
+import "testing"
+
+func TestMockLogger(t *testing.T) {
+	var l Logger = MockLogger{t: t}
+	l.Info("testing")
+}