+++ pkg/errors/errors_test.go
@@ -3,6 +3,7 @@
 package errors
 
 import (
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -22,4 +23,13 @@ func TestErrorUnexpectedStatusCode(t *testing.T) {
 	e := NewUnexpectedStatusCode(99, "wat")
 
 	assert.Equal(t, "99 response returned: wat", e.Error())
+}
+
+func TestErrorUnauthorized(t *testing.T) {
+	t.Parallel()
+
+	e := NewUnauthorizedError()
+
+	assert.Equal(t, 401, e.statusCode)
+	assert.True(t, strings.Contains(e.Error(), "Invalid credentials provided"))
 }